Added conflict uninstall function

+++ src/depsolver/Result.java
@@ -15,7 +15,7 @@ import java.util.ArrayList;
 public class Result {
     public ArrayList<String> result = new ArrayList<String>();
     public Manager newState;
-    public int weight = 0;
+    public double weight = 0;
     
     public Result(Manager curState){
         this.newState = curState;
